requirements.txt
requests>=2.31.0
pdfminer.six>=20220524
PyPDF2>=3.0.0
cryptography>=41.0.0
Pillow>=10.0.0
sv-ttk>=2.5.5
tkinterweb>=3.23.7
tkhtmlview>=0.3.0
